Move creating the AccordionList into its own hook
